Fetch data from local JSON files to fix build errors
