+++ apps/bestofjs-nextjs/src/server/api-utils.tsx
@@ -4,6 +4,12 @@ import { shuffle } from "@/helpers/shuffle";
 
 export const FETCH_ALL_PROJECTS_URL = "https://bestofjs-static-api.vercel.app";
 
+export type RawData = {
+  projects: BestOfJS.RawProject[];
+  tags: BestOfJS.RawTag[];
+  date: string;
+};
+
 export type Data = {
   projectCollection: BestOfJS.RawProject[];
   featuredProjectIds: BestOfJS.Project["slug"][];
