+++ apps/bestofjs-nextjs/src/server/api.ts
@@ -1,0 +1,1 @@
+export * from "./api-local-json";
