+++ apps/bestofjs-nextjs/src/server/create-api.tsx
@@ -5,24 +5,15 @@ import {
   APIContext,
   Data,
   FETCH_ALL_PROJECTS_URL,
+  RawData,
   getFeaturedRandomList,
   getProjectId,
   getTagsByKey,
   populateProject,
 } from "./api-utils";
 
-type RawData = {
-  projects: BestOfJS.RawProject[];
-  tags: BestOfJS.RawTag[];
-  date: string;
-};
-
-export function createAPI() {
+export function createAPI(fetchProjectData: () => Promise<RawData>) {
   async function getData() {
-    return await fetchData();
-  }
-
-  async function fetchData() {
     const { projects, tags: rawTags, date } = await fetchProjectData();
     const tagsByKey = getTagsByKey(rawTags, projects);
     const projectsBySlug: Data["projectsBySlug"] = {};
@@ -55,28 +46,4 @@ export function createAPI() {
     tags: tagsAPI,
     hallOfFame: hallOfFameAPI,
   };
-}
-
-export const api = createAPI();
-
-async function fetchProjectData(): Promise<RawData> {
-  try {
-    const data = await fetchDataFromRemoteJSON();
-    return data as RawData;
-  } catch (error) {
-    console.error("Unable to fetch data!", (error as Error).message);
-    throw error;
-  }
-}
-
-function fetchDataFromRemoteJSON() {
-  const url = FETCH_ALL_PROJECTS_URL + `/projects.json`;
-  console.log(`Fetching JSON data from ${url}`);
-  const options = {
-    next: {
-      revalidate: 60 * 60 * 24, // Revalidate every day
-      tags: ["all-projects"],
-    },
-  };
-  return fetch(url, options).then((res) => res.json());
 }
